move ContextMenu to its own file
